Sync transaction amounts with account balances automatically
Now when you add, edit, or delete a transaction, the linked account's balance updates immediately — no more manually editing wallet/bank balances. Added account selector to the transaction form so you can choose which account (Wallet or Bank) a transaction affects.

+++ src/components/wallet/WalletBalance.jsx
@@ -19,7 +19,7 @@ export default function WalletBalance() {
   if (!wallet || !online) return null
 
   const handleCashSave = () => {
-    const num = parseInt(cashAmount) || 0
+    const num = parseFloat(cashAmount) || 0
     setBalance(wallet.id, num)
     setCashEditing(false)
     setCashSaved(true)
@@ -27,7 +27,7 @@ export default function WalletBalance() {
   }
 
   const handleOnlineSave = () => {
-    const num = parseInt(onlineAmount) || 0
+    const num = parseFloat(onlineAmount) || 0
     setBalance(online.id, num)
     setOnlineEditing(false)
     setOnlineSaved(true)
